chore: rename card to memo
